// app/_components/enums/permissions.ts
// constants/permissions.ts
export enum Permission {
  DASHBOARD = "DASHBOARD",
  HOMEWORK = "HOMEWORK",
  TIMETABLE = "TIMETABLE",
  ATTENDANCE = "ATTENDANCE",
  MARKS = "MARKS",
  CHAT = "CHAT",
  FEES = "FEES",
  CERTIFICATES = "CERTIFICATES",
  ADMISSION = "ADMISSION",

  // School Admin
  CLASSES = "CLASSES",
  STUDENTS = "STUDENTS",
  TEACHERS = "TEACHERS",
  TEACHER_LEAVES = "TEACHER_LEAVES",
  WORKSHOPS = "WORKSHOPS",
  NEWSFEED = "NEWSFEED",
  TC_APPROVALS = "TC_APPROVALS",
  PAYMENTS = "PAYMENTS",
  STUDENT_DETAILS = "STUDENT_DETAILS",
  ANALYSIS = "ANALYSIS",
  EXAMS = "EXAMS",
  NEW_USER = "NEW_USER",
  TEACHER_AUDIT = "TEACHER_AUDIT",

  // Super Admin
  ADD_SCHOOL = "ADD_SCHOOL",
  SCHOOLS = "SCHOOLS",
  FEES_TRANSACTIONS = "FEES_TRANSACTIONS",
  CIRCULARS = "CIRCULARS",
  SETTINGS = "SETTINGS",
  PROFILE = "PROFILE",
  HOME = "HOME",
  STUDENT_LEAVE = "STUDENT_LEAVE",
  LEAVES="LEAVES",
  ANALYTICS = "ANALYTICS"
}